more log

+++ core/src/db/meta/SqliteMetaImpl.cpp
@@ -1302,6 +1302,11 @@ SqliteMetaImpl::CleanUpFilesWithTTL(uint64_t seconds, CleanUpFilter* filter) {
     try {
         server::MetricCollector metric;
 
+        std::vector<int> file_types = {
+            (int)TableFileSchema::TO_DELETE,
+            (int)TableFileSchema::BACKUP,
+        };
+
         // multi-threads call sqlite update may get exception('bad logic', etc), so we add a lock here
         std::lock_guard<std::mutex> meta_lock(meta_mutex_);
 
@@ -1309,21 +1314,23 @@ SqliteMetaImpl::CleanUpFilesWithTTL(uint64_t seconds, CleanUpFilter* filter) {
         auto files = ConnectorPtr->select(columns(&TableFileSchema::id_,
                                                   &TableFileSchema::table_id_,
                                                   &TableFileSchema::file_id_,
+                                                  &TableFileSchema::file_type_,
                                                   &TableFileSchema::date_),
                                           where(
-                                              c(&TableFileSchema::file_type_) ==
-                                              (int)TableFileSchema::TO_DELETE
+                                              in(&TableFileSchema::file_type_, file_types)
                                               and
                                               c(&TableFileSchema::updated_time_)
                                               < now - seconds * US_PS));
 
+        int64_t clean_files = 0;
         auto commited = ConnectorPtr->transaction([&]() mutable {
             TableFileSchema table_file;
             for (auto& file : files) {
                 table_file.id_ = std::get<0>(file);
                 table_file.table_id_ = std::get<1>(file);
                 table_file.file_id_ = std::get<2>(file);
-                table_file.date_ = std::get<3>(file);
+                table_file.file_type_ = std::get<3>(file);
+                table_file.date_ = std::get<4>(file);
 
                 // check if the file can be deleted
                 if (filter && filter->IsIgnored(table_file)) {
@@ -1332,17 +1339,23 @@ SqliteMetaImpl::CleanUpFilesWithTTL(uint64_t seconds, CleanUpFilter* filter) {
                     continue; // ignore this file, don't delete it
                 }
 
-                // delete file from meta
-                ConnectorPtr->remove<TableFileSchema>(table_file.id_);
-
-                // delete file from disk storage
-                utils::DeleteTableFilePath(options_, table_file);
-
-                // erase from cache
+                // erase from cache, must do this before file deleted,
+                // because GetTableFilePath won't able to generate file path after the file is deleted
+                utils::GetTableFilePath(options_, table_file);
                 server::CommonUtil::EraseFromCache(table_file.location_);
 
-                ENGINE_LOG_DEBUG << "Removing file id:" << table_file.file_id_ << " location:" << table_file.location_;
-                table_ids.insert(table_file.table_id_);
+                if (table_file.file_type_ == (int)TableFileSchema::TO_DELETE) {
+                    // delete file from meta
+                    ConnectorPtr->remove<TableFileSchema>(table_file.id_);
+
+                    // delete file from disk storage
+                    utils::DeleteTableFilePath(options_, table_file);
+
+                    ENGINE_LOG_DEBUG << "Removing file id:" << table_file.file_id_ << " location:" << table_file.location_;
+                    table_ids.insert(table_file.table_id_);
+                }
+
+                clean_files++;
             }
             return true;
         });
@@ -1351,8 +1364,8 @@ SqliteMetaImpl::CleanUpFilesWithTTL(uint64_t seconds, CleanUpFilter* filter) {
             return HandleException("CleanUpFilesWithTTL error: sqlite transaction failed");
         }
 
-        if (files.size() > 0) {
-            ENGINE_LOG_DEBUG << "Clean " << files.size() << " files deleted in " << seconds << " seconds";
+        if (clean_files > 0) {
+            ENGINE_LOG_DEBUG << "Clean " << clean_files << " files deleted in " << seconds << " seconds";
         }
     } catch (std::exception& e) {
         return HandleException("Encounter exception when clean table files", e.what());
